+++ test/rocprim/test_device_segmented_scan.cpp
@@ -521,7 +521,10 @@ TYPED_TEST(RocprimDeviceSegmentedScan, InclusiveScanUsingHeadFlags)
             );
             HIP_CHECK(hipDeviceSynchronize());
 
-            ASSERT_NO_FATAL_FAILURE(test_utils::assert_near(output, expected, test_utils::precision_threshold<output_type>::percentage));
+            float multiplier = 1;
+            if(std::is_same<scan_op_type,rocprim::plus<float>>::value)
+                multiplier = 100;
+            ASSERT_NO_FATAL_FAILURE(test_utils::assert_near(output, expected, multiplier*test_utils::precision_threshold<output_type>::percentage));
 
             HIP_CHECK(hipFree(d_temp_storage));
             HIP_CHECK(hipFree(d_input));
@@ -700,8 +703,10 @@ TYPED_TEST(RocprimDeviceSegmentedScan, ExclusiveScanUsingHeadFlags)
                 )
             );
             HIP_CHECK(hipDeviceSynchronize());
-
-            ASSERT_NO_FATAL_FAILURE(test_utils::assert_near(output, expected, test_utils::precision_threshold<output_type>::percentage));
+            float multiplier = 1;
+            if(std::is_same<scan_op_type,rocprim::plus<float>>::value)
+                multiplier = 100;
+            ASSERT_NO_FATAL_FAILURE(test_utils::assert_near(output, expected, multiplier*test_utils::precision_threshold<output_type>::percentage));
 
             HIP_CHECK(hipFree(d_temp_storage));
             HIP_CHECK(hipFree(d_input));