+++ test/rocprim/test_utils.hpp
@@ -50,19 +50,19 @@ static constexpr uint32_t random_data_generation_repeat_strides = 4;
 template<class T>
 struct precision_threshold
 {
-    static constexpr float percentage = 0.01f;
+    static constexpr float percentage = 0.0002f;
 };
 
 template<>
 struct precision_threshold<rocprim::half>
 {
-    static constexpr float percentage = 0.075f;
+    static constexpr float percentage = 0.01f;
 };
 
 template<>
 struct precision_threshold<rocprim::bfloat16>
 {
-    static constexpr float percentage = 0.075f;
+    static constexpr float percentage = 0.02f;
 };
 
 // Support half operators on host side
@@ -455,23 +455,31 @@ struct select_not_equal_to_operator<::rocprim::bfloat16>
     typedef bfloat16_not_equal_to type;
 };
 
-// Plus to operator selector
+/* Plus to operator selector for host-side
+ * On host-side we use `double` as accumulator and `rocprim::plus<double>` as operator
+ * for bfloat16 and half types. This is because additions of floating-point types are not
+ * associative. This would result in wrong output rather quickly for reductions and scan-algorithms
+ * on host-side for bfloat16 and half because of their low-precision.
+ */
 template<typename T>
-struct select_plus_operator
+struct select_plus_operator_host
 {
     typedef ::rocprim::plus<T> type;
+    typedef T acc_type;
 };
 
 template<>
-struct select_plus_operator<::rocprim::half>
+struct select_plus_operator_host<::rocprim::half>
 {
-    typedef half_plus type;
+    typedef ::rocprim::plus<double> type;
+    typedef double acc_type;
 };
 
 template<>
-struct select_plus_operator<::rocprim::bfloat16>
+struct select_plus_operator_host<::rocprim::bfloat16>
 {
-    typedef bfloat16_plus type;
+    typedef ::rocprim::plus<double> type;
+    typedef double acc_type;
 };
 
 // Minimum to operator selector
@@ -643,13 +651,13 @@ template<class InputIt, class T,
                      , bool> = true>
 constexpr T host_reduce(InputIt first, InputIt last, rocprim::plus<T>)
 {
-    using U = double;
+    using accumulator_type = double;
     // Calculate expected results on host
-    U expected = U(0);
-    rocprim::plus<U> bin_op;
+    accumulator_type expected = accumulator_type(0);
+    rocprim::plus<accumulator_type> bin_op;
     for(InputIt it = first; it != last; it++)
     {
-        expected = bin_op(expected, static_cast<U>(*it));
+        expected = bin_op(expected, static_cast<accumulator_type>(*it));
     }
     return static_cast<T>(expected);
 }
@@ -661,9 +669,9 @@ template<class InputIt, class T,
                            , bool> = true>
 constexpr T host_reduce(InputIt first, InputIt last, rocprim::plus<T> op)
 {
-    using U = T;
+    using acc_type = T;
     // Calculate expected results on host
-    U expected = U(0);
+    acc_type expected = acc_type(0);
     for(InputIt it = first; it != last; it++)
     {
         expected = op(expected, *it);
@@ -671,42 +679,96 @@ constexpr T host_reduce(InputIt first, InputIt last, rocprim::plus<T> op)
     return expected;
 }
 
+template<class InputIt, class OutputIt, class BinaryOperation, class acc_type>
+OutputIt host_inclusive_scan_impl(InputIt first, InputIt last,
+                             OutputIt d_first, BinaryOperation op, acc_type)
+{
+    if (first == last) return d_first;
+
+    acc_type sum = *first;
+    *d_first = sum;
+
+    while (++first != last) {
+        sum = op(sum, *first);
+        *++d_first = sum;
+    }
+    return ++d_first;
+}
+
+template<class InputIt, class OutputIt, class BinaryOperation>
+OutputIt host_inclusive_scan(InputIt first, InputIt last,
+                             OutputIt d_first, BinaryOperation op)
+{
+    using acc_type = typename std::iterator_traits<InputIt>::value_type;
+    return host_inclusive_scan_impl(first, last, d_first, op, acc_type{});
+}
+
+template<class InputIt, class OutputIt, class T,
+          std::enable_if_t<std::is_same<typename std::iterator_traits<InputIt>::value_type, rocprim::bfloat16>::value ||
+                           std::is_same<typename std::iterator_traits<InputIt>::value_type, rocprim::half>::value ||
+                           std::is_same<typename std::iterator_traits<InputIt>::value_type, float>::value
+                           , bool> = true>
+OutputIt host_inclusive_scan(InputIt first, InputIt last,
+                             OutputIt d_first, rocprim::plus<T>)
+{
+    using acc_type = double;
+    return host_inclusive_scan_impl(first, last, d_first, rocprim::plus<acc_type>(), acc_type{});
+}
+
+template<class InputIt, class T, class OutputIt, class BinaryOperation, class acc_type>
+OutputIt host_exclusive_scan_impl(InputIt first, InputIt last,
+                             T initial_value, OutputIt d_first,
+                             BinaryOperation op, acc_type)
+{
+    if (first == last) return d_first;
+
+    acc_type sum = initial_value;
+    *d_first = initial_value;
+
+    while ((first+1) != last)
+    {
+        sum = op(sum, *first);
+        *++d_first = sum;
+        first++;
+    }
+    return ++d_first;
+}
+
 template<class InputIt, class T, class OutputIt, class BinaryOperation>
 OutputIt host_exclusive_scan(InputIt first, InputIt last,
                              T initial_value, OutputIt d_first,
                              BinaryOperation op)
 {
-    using result_type = T;
-
-    if (first == last) return d_first;
-
-    result_type sum = initial_value;
-    *d_first = initial_value;
-
-    while ((first+1) != last)
-    {
-       sum = op(sum, *first);
-       *++d_first = sum;
-       first++;
-    }
-    return ++d_first;
+    using acc_type = typename std::iterator_traits<InputIt>::value_type;
+    return host_exclusive_scan_impl(first, last, initial_value, d_first, op, acc_type{});
 }
 
-template<class InputIt, class KeyIt, class T, class OutputIt, class BinaryOperation, class KeyCompare>
-OutputIt host_exclusive_scan_by_key(InputIt first, InputIt last, KeyIt k_first,
-                                    T initial_value, OutputIt d_first,
-                                    BinaryOperation op, KeyCompare key_compare_op)
+template<class InputIt, class T, class OutputIt, class U,
+          std::enable_if_t<std::is_same<typename std::iterator_traits<InputIt>::value_type, rocprim::bfloat16>::value ||
+                               std::is_same<typename std::iterator_traits<InputIt>::value_type, rocprim::half>::value ||
+                               std::is_same<typename std::iterator_traits<InputIt>::value_type, float>::value
+                           , bool> = true>
+OutputIt host_exclusive_scan(InputIt first, InputIt last,
+                             T initial_value, OutputIt d_first,
+                             rocprim::plus<U>)
 {
-    using result_type = T;
+    using acc_type = double;
+    return host_exclusive_scan_impl(first, last, initial_value, d_first, rocprim::plus<acc_type>(), acc_type{});
+}
 
+template<class InputIt, class KeyIt, class T, class OutputIt, class BinaryOperation, class KeyCompare, class acc_type>
+OutputIt host_exclusive_scan_by_key_impl(InputIt first, InputIt last, KeyIt k_first,
+                                         T initial_value, OutputIt d_first,
+                                         BinaryOperation op, KeyCompare key_compare_op, acc_type)
+{
     if (first == last) return d_first;
 
-    result_type sum = initial_value;
+    acc_type sum = initial_value;
     *d_first = initial_value;
 
     while ((first+1) != last)
     {
-        if(key_compare_op(*k_first, *(k_first+1)))
+        if(key_compare_op(*k_first, *++k_first))
         {
             sum = op(sum, *first);
         }
@@ -714,11 +776,78 @@ OutputIt host_exclusive_scan_by_key(InputIt first, InputIt last, KeyIt k_first,
         {
             sum = initial_value;
         }
-	k_first++;
         *++d_first = sum;
         first++;
     }
     return ++d_first;
+}
+template<class InputIt, class KeyIt, class T, class OutputIt, class BinaryOperation, class KeyCompare>
+OutputIt host_exclusive_scan_by_key(InputIt first, InputIt last, KeyIt k_first,
+                                    T initial_value, OutputIt d_first,
+                                    BinaryOperation op, KeyCompare key_compare_op)
+{
+    using acc_type = typename std::iterator_traits<InputIt>::value_type;
+    return host_exclusive_scan_by_key_impl(first, last, k_first, initial_value, d_first, op, key_compare_op, acc_type{});
+}
+
+template<class InputIt, class KeyIt, class T, class OutputIt, class U, class KeyCompare,
+          std::enable_if_t<std::is_same<typename std::iterator_traits<InputIt>::value_type, rocprim::bfloat16>::value ||
+                               std::is_same<typename std::iterator_traits<InputIt>::value_type, rocprim::half>::value ||
+                               std::is_same<typename std::iterator_traits<InputIt>::value_type, float>::value
+                           , bool> = true>
+OutputIt host_exclusive_scan_by_key(InputIt first, InputIt last, KeyIt k_first,
+                                    T initial_value, OutputIt d_first,
+                                    rocprim::plus<U>, KeyCompare key_compare_op)
+{
+    using acc_type = double;
+    return host_exclusive_scan_by_key_impl(first, last, k_first, initial_value, d_first, rocprim::plus<acc_type>(), key_compare_op, acc_type{});
+}
+
+
+template<class InputIt, class KeyIt, class OutputIt, class BinaryOperation, class KeyCompare, class acc_type>
+OutputIt host_inclusive_scan_by_key_impl(InputIt first, InputIt last, KeyIt k_first,
+                                         OutputIt d_first,
+                                         BinaryOperation op, KeyCompare key_compare_op, acc_type)
+{
+    if (first == last) return d_first;
+
+    acc_type sum = *first;
+    *d_first = sum;
+
+    while (++first != last)
+    {
+        if(key_compare_op(*k_first, *++k_first))
+        {
+            sum = op(sum, *first);
+        }
+        else
+        {
+            sum = *first;
+        }
+        *++d_first = sum;
+    }
+    return ++d_first;
+}
+template<class InputIt, class KeyIt, class OutputIt, class BinaryOperation, class KeyCompare>
+OutputIt host_inclusive_scan_by_key(InputIt first, InputIt last, KeyIt k_first,
+                                    OutputIt d_first,
+                                    BinaryOperation op, KeyCompare key_compare_op)
+{
+    using acc_type = typename std::iterator_traits<InputIt>::value_type;
+    return host_inclusive_scan_by_key_impl(first, last, k_first, d_first, op, key_compare_op, acc_type{});
+}
+
+template<class InputIt, class KeyIt, class OutputIt, class U, class KeyCompare,
+          std::enable_if_t<std::is_same<typename std::iterator_traits<InputIt>::value_type, rocprim::bfloat16>::value ||
+                               std::is_same<typename std::iterator_traits<InputIt>::value_type, rocprim::half>::value ||
+                               std::is_same<typename std::iterator_traits<InputIt>::value_type, float>::value
+                           , bool> = true>
+OutputIt host_inclusive_scan_by_key(InputIt first, InputIt last, KeyIt k_first,
+                                    OutputIt d_first,
+                                    rocprim::plus<U>, KeyCompare key_compare_op)
+{
+    using acc_type = double;
+    return host_inclusive_scan_by_key_impl(first, last, k_first, d_first, rocprim::plus<acc_type>(), key_compare_op, acc_type{});
 }
 
 inline
@@ -828,6 +957,13 @@ struct custom_test_type
         return !(*this == other);
     }
 };
+
+template<class T>
+struct precision_threshold<custom_test_type<T>>
+{
+    static constexpr float percentage = 0.01f;
+};
+
 
 //Overload for rocprim::half
 template<>
@@ -1300,7 +1436,7 @@ auto assert_near(const std::vector<T>& result, const std::vector<T>& expected, c
     ASSERT_EQ(result.size(), expected.size());
     for(size_t i = 0; i < result.size(); i++)
     {
-        auto diff = std::max<T>(std::abs(percent * expected[i]), T(percent));
+        auto diff = std::abs(percent * expected[i]);
         ASSERT_NEAR(result[i], expected[i], diff) << "where index = " << i;
     }
 }
@@ -1323,7 +1459,7 @@ void assert_near(const std::vector<rocprim::half>& result, const std::vector<roc
     for(size_t i = 0; i < result.size(); i++)
     {
         if(static_cast<float>(result[i])==static_cast<float>(expected[i])) continue;
-        auto diff = std::max<float>(std::abs(percent * static_cast<float>(expected[i])), percent);
+        auto diff = std::abs(percent * static_cast<float>(expected[i]));
         ASSERT_NEAR(static_cast<float>(result[i]), static_cast<float>(expected[i]), diff) << "where index = " << i;
     }
 }
@@ -1334,7 +1470,7 @@ void assert_near(const std::vector<rocprim::bfloat16>& result, const std::vector
     for(size_t i = 0; i < result.size(); i++)
     {
         if(result[i]==expected[i]) continue;
-        auto diff = std::max<float>(std::abs(percent * static_cast<float>(expected[i])), percent);
+        auto diff = std::abs(percent * static_cast<float>(expected[i]));
         ASSERT_NEAR(static_cast<float>(result[i]), static_cast<float>(expected[i]), diff) << "where index = " << i;
     }
 }
@@ -1344,8 +1480,8 @@ void assert_near(const std::vector<custom_test_type<rocprim::half>>& result, con
     ASSERT_EQ(result.size(), expected.size());
     for(size_t i = 0; i < result.size(); i++)
     {
-        auto diff1 = std::max<float>(std::abs(percent * static_cast<float>(expected[i].x)), percent);
-        auto diff2 = std::max<float>(std::abs(percent * static_cast<float>(expected[i].y)), percent);
+        auto diff1 = std::abs(percent * static_cast<float>(expected[i].x));
+        auto diff2 = std::abs(percent * static_cast<float>(expected[i].y));
         if(static_cast<float>(result[i].x)!=static_cast<float>(expected[i].x)) ASSERT_NEAR(static_cast<float>(result[i].x), static_cast<float>(expected[i].x), diff1) << "where index = " << i;
         if(static_cast<float>(result[i].y)!=static_cast<float>(expected[i].y)) ASSERT_NEAR(static_cast<float>(result[i].y), static_cast<float>(expected[i].y), diff2) << "where index = " << i;
     }
@@ -1356,8 +1492,8 @@ void assert_near(const std::vector<custom_test_type<rocprim::bfloat16>>& result,
     ASSERT_EQ(result.size(), expected.size());
     for(size_t i = 0; i < result.size(); i++)
     {
-        auto diff1 = std::max<float>(std::abs(percent * static_cast<float>(expected[i].x)), percent);
-        auto diff2 = std::max<float>(std::abs(percent * static_cast<float>(expected[i].y)), percent);
+        auto diff1 = std::abs(percent * static_cast<float>(expected[i].x));
+        auto diff2 = std::abs(percent * static_cast<float>(expected[i].y));
         if(result[i].x!=expected[i].x) ASSERT_NEAR(static_cast<float>(result[i].x), static_cast<float>(expected[i].x), diff1) << "where index = " << i;
         if(result[i].y!=expected[i].y) ASSERT_NEAR(static_cast<float>(result[i].y), static_cast<float>(expected[i].y), diff2) << "where index = " << i;
     }
@@ -1371,8 +1507,8 @@ auto assert_near(const std::vector<custom_test_type<T>>& result, const std::vect
     ASSERT_EQ(result.size(), expected.size());
     for(size_t i = 0; i < result.size(); i++)
     {
-        auto diff1 = std::max<T>(std::abs(percent * expected[i].x), T(percent));
-        auto diff2 = std::max<T>(std::abs(percent * expected[i].y), T(percent));
+        auto diff1 = std::abs(percent * expected[i].x);
+        auto diff2 = std::abs(percent * expected[i].y);
         ASSERT_NEAR(result[i].x, expected[i].x, diff1) << "where index = " << i;
         ASSERT_NEAR(result[i].y, expected[i].y, diff2) << "where index = " << i;
     }
@@ -1382,7 +1518,7 @@ template<class T>
 auto assert_near(const T& result, const T& expected, const float percent)
     -> typename std::enable_if<std::is_floating_point<T>::value>::type
 {
-    auto diff = std::max<T>(std::abs(percent * expected), T(percent));
+    auto diff = std::abs(percent * expected);
     ASSERT_NEAR(result, expected, diff);
 }
 
@@ -1397,14 +1533,14 @@ auto assert_near(const T& result, const T& expected, const float percent)
 void assert_near(const rocprim::half& result, const rocprim::half& expected, float percent)
 {
     if(static_cast<float>(result)==static_cast<float>(expected)) return;
-    auto diff = std::max<float>(std::abs(percent * static_cast<float>(expected)), percent);
+    auto diff = std::abs(percent * static_cast<float>(expected));
     ASSERT_NEAR(static_cast<float>(result), static_cast<float>(expected), diff);
 }
 
 void assert_near(const rocprim::bfloat16& result, const rocprim::bfloat16& expected, float percent)
 {
     if(result==expected) return;
-    auto diff = std::max<float>(std::abs(percent * static_cast<float>(expected)), percent);
+    auto diff = std::abs(percent * static_cast<float>(expected));
     ASSERT_NEAR(static_cast<float>(result), static_cast<float>(expected), diff);
 }
 
@@ -1412,8 +1548,8 @@ template<class T>
 auto assert_near(const custom_test_type<T>& result, const custom_test_type<T>& expected, const float percent)
     -> typename std::enable_if<std::is_floating_point<T>::value>::type
 {
-    auto diff1 = std::max<T>(std::abs(percent * expected.x), T(percent));
-    auto diff2 = std::max<T>(std::abs(percent * expected.y), T(percent));
+    auto diff1 = std::abs(percent * expected.x);
+    auto diff2 = std::abs(percent * expected.y);
     ASSERT_NEAR(result.x, expected.x, diff1);
     ASSERT_NEAR(result.y, expected.y, diff2);
 }